+++ custom_components/recalbox/__init__.py
@@ -7,10 +7,11 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.components.http import StaticPathConfig
 from .const import DOMAIN
 from .api import RecalboxAPI
-from .intent import async_setup_intents # Pour charger les phrases Assist
+from .intent import setup_intents
 from .frontend import JSModuleRegistration
 from .translations import RecalboxTranslator
 from .custom_sentences_installer import install_sentences
+from .services_installer import install_services
 import os
 import shutil
 import logging
@@ -34,52 +35,17 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     }
 
     # On enregistre les phrases Assist
-    if "intents_registered" not in hass.data[DOMAIN]:
-        await async_setup_intents(hass)
-        hass.data[DOMAIN]["intents_registered"] = True
+    setup_intents(hass)
 
-    # On ajoute "button" à la liste des plateformes
+    # On ajoute notre switch à la liste des plateformes
     await hass.config_entries.async_forward_entry_setups(entry, ["switch"])
 
-
-
-    # services appelés par le JS
-
-    async def handle_shutdown(call):
-        recalbox_entity = findRecalboxEntity(hass, call.data.get("entity_id"))
-        if recalbox_entity: await recalbox_entity.request_shutdown()
-    async def handle_reboot(call):
-        recalbox_entity = findRecalboxEntity(hass, call.data.get("entity_id"))
-        if recalbox_entity: await recalbox_entity.request_reboot()
-    async def handle_screenshot(call):
-        recalbox_entity = findRecalboxEntity(hass, call.data.get("entity_id"))
-        if recalbox_entity: await recalbox_entity.request_screenshot()
-    async def handle_quit_game(call):
-        recalbox_entity = findRecalboxEntity(hass, call.data.get("entity_id"))
-        if recalbox_entity: await recalbox_entity.request_quit_current_game()
-    async def handle_launch_game(call):
-        recalbox_entity = findRecalboxEntity(hass, call.data.get("entity_id"))
-        game = call.data.get("game")
-        console = call.data.get("console")
-        if recalbox_entity:
-            await recalbox_entity.search_and_launch_game_by_name(console, game)
-
-    # Enregistrement du service recalbox.screen
-    hass.services.async_register(DOMAIN, "shutdown", handle_shutdown)
-    hass.services.async_register(DOMAIN, "reboot", handle_reboot)
-    hass.services.async_register(DOMAIN, "screenshot", handle_screenshot)
-    hass.services.async_register(DOMAIN, "quit_game", handle_quit_game)
-    hass.services.async_register(DOMAIN, "launch_game", handle_launch_game)
+    # rengistrement des services Recalbox, utilisés par la partie JS notamment
+    # mais dispo aussi dans HA au global
+    install_services(hass)
 
     return True
 
-
-def findRecalboxEntity(hass: HomeAssistant, entity_id: str):
-    for instance in hass.data[DOMAIN]["instances"].values():
-        entity = instance.get("sensor_entity")
-        if entity and entity.entity_id == entity_id:
-            return entity
-    return None
 
 
 async def async_register_frontend(hass: HomeAssistant) -> None:
@@ -98,7 +64,7 @@ async def async_setup(hass: HomeAssistant, config: dict) -> bool:
     # Etape préliminaire :
     # Installer les phrases Assist automatiquement
     hass.data[DOMAIN]["global"]["needs_restart"]= await hass.async_add_executor_job(
-        install_sentences, hass, DOMAIN
+        install_sentences, hass
     )
 
     # enregistrement du chemin statique pour la custom card